Move get_title_from_url tests from test_services to test_utils
